Added table plugin

+++ domain/issue.go
@@ -3,6 +3,6 @@ package domain
 type Issue struct {
 	ID        string `json:id`
 	Title     string `json:title`
-	UserCount int64  `json:userCount`
+	UserCount int    `json:userCount`
 	Permalink string `json:permalink`
 }
